Cleaned up fork

+++ kern/arch/mips/syscall/syscall.c
@@ -147,7 +147,7 @@ syscall(struct trapframe *tf)
 	/* Process system calls */
 
 	case SYS_execv:
-		err = (int32_t)sys_execv((userptr_t)tf->tf_a0, (userptr_t)tf->tf_a1);
+		err = sys_execv((userptr_t)tf->tf_a0, (userptr_t)tf->tf_a1);
 		break;
 
 	case SYS__exit:
@@ -155,7 +155,7 @@ syscall(struct trapframe *tf)
 		break;
 
 	case SYS_fork:
-		err = (int32_t)sys_fork(tf, (pid_t *)&retval);
+		err = sys_fork(tf, (pid_t *)&retval);
 		break;
 
 	case SYS_getpid:
@@ -163,7 +163,7 @@ syscall(struct trapframe *tf)
 		break;
 
 	case SYS_waitpid:
-		err = (int32_t)sys_waitpid((pid_t)tf->tf_a0, (int*)tf->tf_a1, (int)tf->tf_a2, (pid_t *)&retval);
+		err = sys_waitpid((pid_t)tf->tf_a0, (int*)tf->tf_a1, (int)tf->tf_a2, (pid_t *)&retval);
 		break;
 
 	default:
